[M140] CacheStorage: Support large file on Cache Storage API
Original change's description:
> CacheStorage: Support large file on Cache Storage API
> 
> This is a last cl required for supporting 2GB+ file.
> 
> Bug: 391398191
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6705024
> Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
> Auto-Submit: Eriko Kurimoto <elkurin@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1500000}

Bug: 438100035,391398191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6849492
Auto-Submit: Chrome Cherry Picker <chrome-cherry-picker@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/7339@{#660}
Cr-Branched-From: 27be8b77710f4405fdfeb4ee946fcabb0f6c92b2-refs/heads/main@{#1496484}

+++ content/browser/cache_storage/cache_storage_blob_to_disk_cache.cc
@@ -83,8 +83,7 @@ void CacheStorageBlobToDiskCache::OnComplete(int32_t status,
   received_on_complete_ = true;
   expected_total_size_ = data_length;
   if (data_pipe_closed_) {
-    RunCallback(static_cast<uint64_t>(cache_entry_offset_) ==
-                expected_total_size_);
+    RunCallback(cache_entry_offset_ == expected_total_size_);
   }
 }
 
@@ -129,8 +128,7 @@ void CacheStorageBlobToDiskCache::OnDataPipeReadable(MojoResult unused) {
     // Done reading, but only signal success if OnComplete has also been called.
     data_pipe_closed_ = true;
     if (received_on_complete_) {
-      RunCallback(static_cast<uint64_t>(cache_entry_offset_) ==
-                  expected_total_size_);
+      RunCallback(cache_entry_offset_ == expected_total_size_);
     }
     return;
   }
@@ -149,8 +147,9 @@ void CacheStorageBlobToDiskCache::OnDataPipeReadable(MojoResult unused) {
                      weak_ptr_factory_.GetWeakPtr(), bytes_to_read);
 
   int rv = entry_->WriteData(
-      disk_cache_body_index_, cache_entry_offset_, buffer.get(), bytes_to_read,
-      std::move(cache_write_callback), true /* truncate */);
+      disk_cache_body_index_, base::checked_cast<int64_t>(cache_entry_offset_),
+      buffer.get(), bytes_to_read, std::move(cache_write_callback),
+      true /* truncate */);
 
   if (rv != net::ERR_IO_PENDING) {
     CacheStorageBlobToDiskCache::DidWriteDataToEntry(bytes_to_read, rv);
